reduces oracle pushes to 10k & top 30 yield
