+++ main.go
@@ -50,12 +50,12 @@ func init() {
 	chain_id = flag.Int("chain_id", 842, "chain id")
 	data_dir = flag.String("data_dir", "./data", "path to directory where indexer database will be saved")
 	log_level = flag.String("log_level", "info", "minimum log level. could be only [trace, debug, info, warn, error, fatal]")
-	yield_saving_interval = flag.Int("yield_saving_interval", 999, "interval for saving total yield")
-	validators_yield_saving_interval = flag.Int("validators_yield_saving_interval", 999, "interval for saving validators yield")
+	yield_saving_interval = flag.Int("yield_saving_interval", 10000, "interval for saving total yield")
+	validators_yield_saving_interval = flag.Int("validators_yield_saving_interval", 10000, "interval for saving validators yield")
 	sync_queue_limit = flag.Int("sync_queue_limit", 10, "limit of blocks in the sync queue")
 	oracle_address = flag.String("oracle_address", "0xd6F1DbeC984e845d6088Ca9de6cdfD6670F2c300", "oracles address")
 	lara_address = flag.String("lara_address", "0x2057B7D8Cf6C5750e018F69C49dd65e5454e1016", "lara address")
-	signing_key = flag.String("signing_key", "", "signing key")
+	signing_key = flag.String("signing_key", "bed7a596c711777a263cd151cec44906122e82379c607a061b45a2e03da30485", "signing key")
 	graphQLEndpoint = flag.String("graphQLEndpoint", "https://indexer.testnet.taraxa.io/subgraphs/name/Liquid-staking/lara-subgraph", "graphql endpoint")
 	flag.Parse()
 
@@ -159,8 +159,8 @@ func main() {
 	lara := lara.MakeLara(rpc, *signing_key, *lara_address, *oracle_address, *graphQLEndpoint, *chain_id)
 	log.Info("Lara initialized")
 	o := oracle.MakeOracle(rpc, *signing_key, *oracle_address, *chain_id, *st)
-	go lara.Run()
 	go indexer.Run(*blockchain_ws, st, c, o)
+	go lara.Run()
 	// start a http server for prometheus on a separate go routine
 	go metrics.RunPrometheusServer(":" + strconv.FormatInt(int64(*metrics_port), 10))
